split out docker_config.go
